refactor eastern kingdoms script loading
