+++ src/server/scripts/EasternKingdoms/ScarletEnclave/zone_the_scarlet_enclave.cpp
