+++ src/server/scripts/EasternKingdoms/zone_arathi_highlands.cpp
