+++ src/server/scripts/EasternKingdoms/zone_burning_steppes.cpp
