+++ src/server/scripts/EasternKingdoms/zone_eastern_plaguelands.cpp
