+++ src/server/scripts/EasternKingdoms/zone_eversong_woods.cpp
