+++ src/server/scripts/EasternKingdoms/zone_ghostlands.cpp
