+++ src/server/scripts/EasternKingdoms/zone_hinterlands.cpp
