+++ src/server/scripts/EasternKingdoms/zone_isle_of_queldanas.cpp
