+++ src/server/scripts/EasternKingdoms/zone_loch_modan.cpp
