+++ src/server/scripts/EasternKingdoms/zone_redridge_mountains.cpp
