+++ src/server/scripts/EasternKingdoms/zone_silverpine_forest.cpp
