+++ src/server/scripts/EasternKingdoms/zone_stranglethorn_vale.cpp
