+++ src/server/scripts/EasternKingdoms/zone_tirisfal_glades.cpp
