+++ src/server/scripts/EasternKingdoms/zone_western_plaguelands.cpp
