+++ src/server/scripts/Maelstrom/LostIsles/kezan.cpp
