+++ src/server/scripts/Maelstrom/TheStonecore/boss_high_priestess_azil.cpp
