+++ src/server/scripts/Maelstrom/TheStonecore/boss_ozruk.cpp
