+++ src/server/scripts/Maelstrom/TheStonecore/boss_slabhide.cpp
