+++ src/server/scripts/Maelstrom/TheStonecore/the_stonecore.h
